feat: restructure pages and add dashboard
- Add new dashboard.vue page as main entry point
- Refactor index.vue with enhanced layout and functionality
- Update analytics.vue and reports.vue with improved UI
- Remove old home.vue page (replaced by dashboard)
- Update default layout navigation structure
